+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_compass/flutter_compass.dart';
+import 'package:flutter_compass_v2/flutter_compass_v2.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 void main() => runApp(MyApp());
